TokenAlgebra Example: ValueSet
This PR provides a new module called Ledger.TokenAlgebra.ValueSet which uses the new functionality in TotalMap to define a token algebra using the Theory record of the Axiom.Set module.

To implement a token algebra using Set and modeling the value type as a total function on a given X : Set A I needed to add some functionality to TotalMap (and similar functionality to a new module, called TotalMapOn which restricts maps to the set X, but I have since abandoned the approach using TotalMapOn as I couldn't get it to work without adding a dependently typed version of the replacement axiom of the Set module, and we don't want to add axioms at this point for a number of reasons; this is why the comments below are crossed out).

The new functionality in TotalMap is part of a separate PR which should be merged before this one.

+  Update src/Axiom/Set/TotalMap.agda
   Co-authored-by: Andre Knispel <andre.knispel@gmx.de>

+  Update src/Ledger/TokenAlgebra/ValueSet.lagda
   Co-authored-by: Andre Knispel <andre.knispel@gmx.de>

+  incorporate @WhatisRT's requested changes