Type classes for common structures
This addresses issue #34 for the special case of inequalities.

The main goal of this PR is to allow us to use a single operation
symbol, `_<_`, to denote the less than relation on the types `ℤ`, `ℕ`,
and the `Slot` and `Epoch` types occurring in fields of the
`EpochStructure` record type.

In fact, I could only find one instance of a special, decorated
less than symbol---namely, the symbol `<ˢ` occurrs on line 97 of
the `PPUp.lagda` file.

The new `HasStrictTotalOrder` module, and the instance for
`HasStrictTotalOrder Slot _≡_` in the `HasStrictTotalOrder/Instance`
module, allows us to write `<` on line 97 of `PPUp.lagda` (instead of `<ˢ`).

Some purely cosmetic/no-op improvements were removed from this PR, but
these should probably be included in a future PR.  In particular,

+  a couple of modules we touch here import the entire `Algebras` module
   in cases where we only really need `open import Algebras.Bundles using
   (Semiring)`.

+  Another minor cosmetic improvement would be to change the lines in
   equational reasoning proofs near the bottom of `Utxo.Properties` that
   begin with `ℤ.+_`; these can all begin with `ℤ.+` instead.