Type classes for common structures
This addresses issue #34 for the special case of addition in the
`Ledger.Utxo.Properties` module.

The main goal of this PR is to allow us to use a single operation
symbol, `_+_`, to denote addition over each of the types `ℤ`, `ℕ`,
and `TokenAlgebra.Value`.

Most of the changes in this PR---including the clean-up of proofs and
the removal of some uses of tactics---were made in an effort to support
this operator overloading (via implicit instances), including

+  a large number of changes are proposed that significantly clean up and simplify
   the presentation and the proofs in the `Utxo.Properties` module

+  eliminate unused code (imports, private variables, and unnecessary implicit arguments)

+  use of tactics in equational reasoning proofs were eliminated because I was
   unable to get the tactic proofs to type-check when using the new implicit
   instances for types with addition

+  change some instances of `+` to their fully qualified versions so that
   the proofs involving tactics can remain unchanged.

- put `HasAdd` record in its own module under `Interface`
- add instances throughout the library and replace annotate add symbols with `_+_`
- fix proof of `mostStakeDRepDist-∅`; it requires annotated add symbol (`_+ℕ_`)