add type class for subtraction
This addresses issue #34 for the special case of subtraction in the
`GovernanceActions`, `Utxo.Properties`, and `Tactic/DeriveComp` modules.

The main goal of this PR is to allow us to use a single operation symbol, _-_,
to denote subtraction over each of the types ℤ, ℕ, and Coin (which is currently
just an alias for ℕ).