Type classes for common structures: preorders and partial orders
This addresses issue #34 for the special cases of preorders and partial orders.

The goal of this PR is to allow us to use

+ a single relation symbol ≤, to denote the "less or equal" relation on each of the types ℤ, ℕ, and Slot---on which ≤ denotes a partial order---and on the Epoch type, on which ≤` denotes a preorder.

+ a single relation symbol < to denote the "strictly less than" relation on each of the types ℤ, ℕ, and Slot.

---

We have temporarily removed some infered symbols to type-check.

We have changed from using raw partial orders to preorders and partial orders with laws.

(we probably need a preorder type class if we want to use the leq order symbol
on the Epoch type without annotations, since that order is not antisymmetric)