Fix table rendering in Gov.Actions module
The longtable and even the simple table environment causes pandoc to put the markdown version of the table inside a <div></div> block and this apparently prevents mkdocs from rendering the table correctly, despite the fact that we are using the markdown extension `md_in_html`.  There ought to be a way to fix this, but until a better way is found, a workaround is to simply put the table in a tabular environment and omit the table environment altogether.

Note: for some reason pandoc silently deletes the footnote in the table.

To resolve both the table rendering issue and the lost footnote, I suggest we hand-tweak the `Actions.lagda.md` file and commit it to the repo.