add a means of linking to original source files
Because our pipeline generates pure markdown (`.md`) files from four
different kinds of files (`.tex`, `.agda`, `.lagda`, `.lagda.md`), the
ability to refer back to the original file name is not automatic.  This
commit adds a bunch of Python functionality that is needed to basically
just insert a yaml header in the final markdown file that points to the
source of that file.  At this point, it gets it right in most (but not
all) cases.  Example exception: `Notation.md` incorrectly points to

https://github.com/IntersectMBO/formal-ledger-specifications/blob/master/src/Notation.lagda.md

It should point to

https://github.com/IntersectMBO/formal-ledger-specifications/blob/master/build-tools/static/md/common/src/Notation.md