revert CERTS rule to do CERTBASE last
This is a first step to refactoring the CERTS rule.

Ultimately we will deprecate the "CERTBASE" rule and replace it
with "preCerts" and "postCerts" functions that will run before and after
the CERTS sts.  Some of the things that currently happen in CERTBASE may
be moved into the LEDGER rule.

revert CERTS rule in conformance version